k2sbridge/requirements.txt
cached-property==1.3.0
docker-compose==1.6.2
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
enum34==1.1.2
functools32==3.2.3.post2
jsonschema==2.5.1
kafka==1.0.2
PyYAML==3.11
requests==2.7.0
six==1.10.0
stomp.py==4.1.9
texttable==0.8.4
websocket-client==0.35.0
wheel==0.24.0
